+++ legacy/src/layout/questionnaire-edit/questionnaire-edit.jsx
@@ -3,6 +3,7 @@ import { useContext } from 'react';
 import PropTypes from 'prop-types';
 
 import { AuthContext } from '@/auth/context';
+import { getStampFromEmail } from '@/utils/utils';
 
 import { COMPONENT_TYPE } from '../../constants/pogues-constants';
 import { Component as ComponentFactory } from '../../model';
@@ -93,7 +94,8 @@ function QuestionnaireNew({
   activeCodeListsById,
 }) {
   const { decodedIdToken } = useContext(AuthContext);
-  const stamp = decodedIdToken.timbre;
+  const stamp =
+    decodedIdToken.timbre ?? getStampFromEmail(decodedIdToken.email);
   const validate = (setErrorsAction) => (values) =>
     validateQuestionnaireForm(values, setErrorsAction);
   // Initial values